requirements.txt
# Django & Core
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-filter==24.2
python-decouple==3.8
psycopg2-binary==2.9.9
redis==5.0.4
django-redis==5.4.0

# Celery & Tasks
celery==5.3.6
django-celery-beat==2.6.0
django-celery-results==2.5.1

# Utils
Pillow==10.3.0
qrcode[pil]==7.4.2
gunicorn==22.0.0
whitenoise==6.6.0

# Testing
pytest==8.2.0
pytest-django==4.8.0
factory-boy==3.3.0
faker==25.0.1
